requirements.txt
absl-py==0.15.0
asgiref==3.4.1
astunparse==1.6.3
async-generator==1.10
attrs==21.2.0
Automat==20.2.0
beautifulsoup4==4.10.0
botocore==1.21.42
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.5
clang==5.0
cloudinary==1.28.0
cloudpickle==2.0.0
colorama==0.4.4
configparser==5.2.0
constantly==15.1.0
crayons==0.4.0
cryptography==3.4.8
cssselect==1.1.0
cycler==0.11.0
defusedxml==0.7.1
dj-database-url==0.5.0
dj3-cloudinary-storage==0.0.6
Django==4.0.1
django-allauth==0.47.0
django-bootstrap-form==3.4
django-braces==1.14.0
django-ckeditor==6.1.0
django-comments-xtd==2.9.2
django-contrib-comments==2.1.0
django-cors-headers==3.8.0
django-crispy-forms==1.12.0
django-etc==1.3.0
django-filter==2.4.0
django-heroku==0.3.1
django-hitcount==1.3.3
django-ipware==4.0.0
django-js-asset==1.2.2
django-model-utils==4.1.1
django-modeltranslation==0.17.3
django-recaptcha3==0.4.0
django-star-ratings==0.9.1
django-storages==1.11.1
django-summernote==0.8.11.6
django-tinymce==3.4.0
django-tinymce4-lite==1.8.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
docutils==0.17.1
flatbuffers==1.12
fonttools==4.28.4
gast==0.4.0
google-auth==2.1.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.41.1
gunicorn==20.1.0
h11==0.12.0
h2==3.2.0
h5py==3.1.0
heroku==0.1.4
hpack==3.0.0
hyperframe==5.2.0
hyperlink==21.0.0
idna==3.2
incremental==21.3.0
itemadapter==0.4.0
itemloaders==1.0.4
jmespath==0.10.0
joblib==1.1.0
jsmin==3.0.0
keras==2.6.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
lxml==4.7.1
Markdown==3.3.4
matplotlib==3.5.1
numpy==1.19.5
oauthlib==3.1.1
opencv-python==4.5.4.60
opt-einsum==3.3.0
outcome==1.1.0
packaging==21.3
pandas==1.3.4
parsel==1.6.0
Pillow==8.3.2
priority==1.3.0
Protego==0.1.16
protobuf==3.19.1
psycopg2==2.9.1
psycopg2-binary==2.9.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyDispatcher==2.0.5
PyJWT==2.1.0
pyOpenSSL==21.0.0
pyparsing==3.0.6
python-dateutil==1.5
python3-openid==3.2.0
pytz==2021.1
queuelib==1.6.2
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
s3transfer==0.5.0
schedule==1.1.0
scikit-learn==1.0.1
scipy==1.7.1
Scrapy==2.5.1
selenium==4.1.0
service-identity==21.1.0
six==1.15.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.2.1
sqlparse==0.4.2
swapper==1.1.2.post1
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.6.1
tensorflow-estimator==2.6.0
termcolor==1.1.0
threadpoolctl==3.0.0
tqdm==4.62.2
trio==0.19.0
trio-websocket==0.9.2
Twisted==21.7.0
twisted-iocpsupport==1.0.2
typing-extensions==3.7.4.3
tzdata==2021.5
urllib3==1.26.6
w3lib==1.22.0
webdriver-manager==3.5.2
Werkzeug==2.0.2
whitenoise==5.3.0
wrapt==1.12.1
wsproto==1.0.0
yahoofinancials==1.6
zope.interface==5.4.0
